requirements.txt
requests
Flask
oauthlib
pyOpenSSL
pyclamd
pillow